+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* Managing people
 * Rename `classes` resource to `courses`
 
 ### 0.1.1